requirements.txt
Flask
Pillow
pytesseract